게시글 수정 시 이미지 첨부 수정

See merge request s09-webmobile1-sub2/S09P12D202!174

+++ backend/HonJaRang/src/main/java/com/example/honjarang/domain/post/service/PostService.java
@@ -16,6 +16,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.expression.spel.ast.NullLiteral;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -85,26 +86,63 @@ public class PostService {
             throw new InvalidUserException("작성자만 수정할 수 있습니다.");
         }
 
-        // 기존에 게시글에 사진이 있으면 사진을 s3에서 지우고 추가
-        if (post.getPostImage() != "") {
+
+        // 기존에 게시글에 사진이 있었고 사용자가 사진을 첨부했을 때
+        // 기존의 사진을 삭제하고 새로운 사진 추가
+        if (post.getPostImage()!= null && postImage!=null) {
             DeleteObjectRequest request = DeleteObjectRequest.builder()
                     .bucket("honjarang-bucket")
                     .key("postImage/" + post.getPostImage())
                     .build();
             s3Client.deleteObject(request);
 
+             // 사진추가
+            String uuid = UUID.randomUUID().toString();
+            s3Client.putObject(PutObjectRequest.builder()
+                    .bucket("honjarang-bucket")
+                    .key("postImage/" + uuid + postImage.getOriginalFilename())
+                    .acl(ObjectCannedACL.PUBLIC_READ)
+                    .contentType(postImage.getContentType())
+                    .build(), RequestBody.fromInputStream(postImage.getInputStream(), postImage.getSize()));
+
+            String image = uuid + postImage.getOriginalFilename();
+            post.update(postUpdateDto);
+            post.updateImage(image);
+
+        }
+        
+        // 기존에 게시글에 사진이 있었고 사용자가 사진을 첨부하지 않았을 때
+        // 그냥 놔둬야 됨, 아무것도 건들이지 않음
+        // 아예 parameter를 없애야 성공적으로 됨
+        if (post.getPostImage()!= null && postImage==null) {
+            post.update(postUpdateDto);
+            post.updateImage(post.getPostImage());
         }
 
-        String uuid = UUID.randomUUID().toString();
-        s3Client.putObject(PutObjectRequest.builder()
-                .bucket("honjarang-bucket")
-                .key("postImage/" + uuid + postImage.getOriginalFilename())
-                .acl(ObjectCannedACL.PUBLIC_READ)
-                .contentType(postImage.getContentType())
-                .build(), RequestBody.fromInputStream(postImage.getInputStream(), postImage.getSize()));
 
-        String image = uuid + postImage.getOriginalFilename();
-        post.update(postUpdateDto,image);
+        // 기존에 사진이 없었고 사용자가 사진을 첨부했을 때
+        // 사진 추가와 글 업데이트
+        if (post.getPostImage()== null && postImage!=null) {
+            // 사진추가
+            String uuid = UUID.randomUUID().toString();
+            s3Client.putObject(PutObjectRequest.builder()
+                    .bucket("honjarang-bucket")
+                    .key("postImage/" + uuid + postImage.getOriginalFilename())
+                    .acl(ObjectCannedACL.PUBLIC_READ)
+                    .contentType(postImage.getContentType())
+                    .build(), RequestBody.fromInputStream(postImage.getInputStream(), postImage.getSize()));
+
+            String image = uuid + postImage.getOriginalFilename();
+            post.update(postUpdateDto);
+            post.updateImage(image);
+        }
+
+        // 기존에 사진이 없었고 사용자가 사진을 첨부하지 않았을 때
+        // 그냥 글만 업데이트 하면됨
+        if (post.getPostImage()== null && postImage==null) {
+            post.update(postUpdateDto);
+            post.updateImage(null);
+        }
 
     }
